pyproject.toml
[project]
name = "mcp-eshipz"
version = "0.1.0"
description = "eShipz tracking MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0"
]
